Move multiple Java class files into various packages
